Change pytest to stop after first failure
I'm not super sure about this but I think it will be friendlier
for the students to not see a huge wall of failed tests.

A possible concern is that only seeing one failure could be
confusing to students, but my hope is that it is easily learnable
that they are working through the failures one by one.